# architecture/api/openapi.yaml
openapi: 3.0.3
info:
  title: GazetaMarketplace — API interna
  version: 1.0.0
  description: |
    Endpoints JSON usados pelas páginas do próprio site (mesma origem, sem CORS).
    As páginas do site público e do painel são renderizadas no servidor (Razor) e não fazem parte deste contrato.

    Contrato de erros: ProblemDetails (RFC 7807) com o campo `code` da tabela de `ARCHITECTURE.md` §8
    (`VALIDATION_ERROR`, `UNAUTHORIZED`, `FORBIDDEN`, `NOT_FOUND`, `CONFLICT`, `RATE_LIMITED`,
    `CEP_SERVICE_UNAVAILABLE`, `INTERNAL_ERROR`). `traceId` sempre presente; é o "código de referência" das telas de erro.

    Autenticação: cookie de sessão do ASP.NET Core Identity (ADR-003). Escritas exigem também o cabeçalho
    `RequestVerificationToken` (antiforgery, NFR-11).
servers:
  - url: /
    description: Mesma origem do site
...
paths:
...
  /api/v1/ads/{adId}/photos:
    post:
      tags: [Fotos]
      summary: Enviar uma foto para um anúncio
      description: |
        Uma foto por requisição. Aceita JPG, PNG, WebP, GIF e HEIC, conferidos pelo conteúdo do arquivo; até 10 MB.
        O servidor grava o original por 30 dias e gera duas versões WebP (1.600 px e 480 px de largura, e no máximo 2.560 px no lado maior) sem metadados.
        O limite do servidor para o corpo é de 11 MB, para que um arquivo de 10 a 11 MB receba a mensagem
        `VALIDATION_ERROR` da aplicação; acima disso, o próprio servidor recusa com 413.
        Limite de fotos pelo grupo de campos da categoria: 20 (padrão), 6 (Serviços), 0 (Vagas de emprego).
        No máximo 30 envios por minuto por usuário (cada arquivo conta); o excesso recebe 429 `RATE_LIMITED` com `Retry-After`.
        Só o autor (em Rascunho ou Rejeitado) ou o Administrador envia; cada envio trava a linha do anúncio, então envios ao mesmo tempo nunca passam do limite.
      operationId: uploadAdPhoto
      security:
        - cookieAuth: []
          antiforgery: []
      parameters:
        - $ref: '#/components/parameters/AdId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Arquivo de imagem
      responses:
        '201':
          description: Foto processada e adicionada ao fim da galeria (a primeira é a capa)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          description: Corpo acima de 11 MB, recusado pelo servidor antes de ler o arquivo
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/ads/{adId}/photos/{photoId}:
    delete:
      tags: [Fotos]
      summary: Remover uma foto do anúncio
      description: Remove a foto e seus arquivos WebP (US-008-S03). Se era a capa, a próxima foto passa a ser a capa.
      operationId: deleteAdPhoto
      security:
        - cookieAuth: []
          antiforgery: []
      parameters:
        - $ref: '#/components/parameters/AdId'
        - $ref: '#/components/parameters/PhotoId'
      responses:
        '204':
          description: Foto removida
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Identity.Application
      description: Cookie de sessão do Identity (HttpOnly, Secure, SameSite=Lax; 30 min sem uso)
    antiforgery:
      type: apiKey
      in: header
      name: RequestVerificationToken
      description: Token antiforgery lido da página (meta `request-verification-token`)
  parameters:
...
    AdId:
      name: adId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        example: 104
...
    PhotoId:
      name: photoId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        example: 877
  schemas:
...
    Photo:
      type: object
      required: [id, sortOrder, url480, url1600, width, height]
      properties:
        id:
          type: integer
          example: 877
        sortOrder:
          type: integer
          description: 0 = capa
          example: 0
        url480:
          type: string
          example: /fotos/104/877-480.webp
        url1600:
          type: string
          example: /fotos/104/877-1600.webp
        width:
          type: integer
          example: 1600
        height:
          type: integer
          example: 1200
...
    ProblemDetails:
      type: object
      description: Erro no formato RFC 7807, com `code` estável (ARCHITECTURE.md §8)
      required: [status, title, code, traceId]
      properties:
        type:
          type: string
          example: https://tools.ietf.org/html/rfc7231#section-6.5.1
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Formato não aceito. Use JPG, PNG, WebP, GIF ou HEIC
        instance:
          type: string
          example: /api/v1/ads/104/photos
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, CEP_SERVICE_UNAVAILABLE, INTERNAL_ERROR]
          example: VALIDATION_ERROR
        traceId:
          type: string
          example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
        errors:
          type: object
          description: Só em VALIDATION_ERROR; mensagens por campo
          additionalProperties:
            type: array
            items:
              type: string
          example:
            file: ['A foto excede o limite de 10 MB']
  responses:
    ValidationError:
      description: Dados inválidos (VALIDATION_ERROR)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Sem sessão válida (UNAUTHORIZED)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            status: 401
            title: Unauthorized
            code: UNAUTHORIZED
            detail: Faça login para continuar
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    Forbidden:
      description: Sem permissão para o anúncio (FORBIDDEN)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            status: 403
            title: Forbidden
            code: FORBIDDEN
            detail: Você não tem permissão para acessar este anúncio
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    NotFound:
      description: Recurso inexistente ou CEP não encontrado (NOT_FOUND)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            status: 404
            title: Not Found
            code: NOT_FOUND
            detail: CEP não encontrado. Confira os números.
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    Conflict:
      description: Limite de fotos atingido, anúncio em situação que não permite a ação, ou edição simultânea (CONFLICT)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            status: 409
            title: Conflict
            code: CONFLICT
            detail: Cada anúncio pode ter no máximo 20 fotos
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    RateLimited:
      description: Muitas requisições (RATE_LIMITED)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    InternalError:
      description: Erro inesperado (INTERNAL_ERROR); detalhes só no log
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            status: 500
            title: Internal Server Error
            code: INTERNAL_ERROR
            detail: Ocorreu um erro inesperado
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01